Update README.md
The goal of this project is to perform a comprehensive Exploratory Data Analysis (EDA) on a dataset with the ultimate aim of preparing the data for machine learning modeling. Through this process, we aim to gain valuable insights into the dataset's structure, patterns, and potential challenges, which will inform the subsequent steps in our machine learning pipeline.